[IMP] mail: unskip "chat with author should be opened after clicking on their im status icon"
