+++ addons/mail/static/tests/new/message/message_tests.js
@@ -1171,3 +1171,35 @@ QUnit.test(
         assert.containsOnce(target, `.o-mail-thread[data-thread-id=${partnerId}]`);
     }
 );
+
+QUnit.test(
+    "chat with author should be opened after clicking on their im status icon",
+    async function (assert) {
+        const pyEnv = await startServer();
+        const [threadId, partnerId] = pyEnv["res.partner"].create([{}, { im_status: "online" }]);
+        pyEnv["res.users"].create({
+            im_status: "online",
+            partner_id: partnerId,
+        });
+        pyEnv["mail.message"].create({
+            author_id: partnerId,
+            body: "not empty",
+            model: "res.partner",
+            res_id: threadId,
+        });
+        const { advanceTime, openFormView } = await start({
+            hasTimeControl: true,
+        });
+        await openFormView({
+            res_id: threadId,
+            res_model: "res.partner",
+        });
+        await afterNextRender(() => advanceTime(50 * 1000)); // next fetch of im_status
+        assert.containsOnce(target, ".o-mail-partner-im-status");
+        assert.hasClass(target.querySelector(".o-mail-partner-im-status"), "cursor-pointer");
+
+        await click(".o-mail-partner-im-status");
+        assert.containsOnce(target, ".o-mail-chat-window");
+        assert.containsOnce(target, `.o-mail-thread[data-thread-id=${threadId}]`);
+    }
+);
